+++ .gitignore
@@ -1,1 +1,2 @@
 mapkit-android-demo-master/
+*/OLD/
