--- API/.gitignore
@@ -1,11 +1,0 @@
-# VS Code Files
-.vscode/
-
-# Python Files
-__pycache__/
-
-# Database
-*.db
-
-# Tokens and etc.
-*.ini
