+++ API/backend/NEW/.gitignore
@@ -1,0 +1,2 @@
+# DB files
+*.db
