+++ API/backend/NEW/config.py
@@ -1,0 +1,4 @@
+token = "ba661e842cfe7b9dce1a5153c6e80d5e"
+map_token = "ccbc665b-e2a1-4c7d-aff9-cd458b1317bb"
+# path_to_db = "./DB/.db"
+path_to_db = "./DB/Toilets.db"
