+++ API/backend/NEW/requirements.txt
@@ -1,0 +1,5 @@
+sqlalchemy
+uvicorn
+fastapi
+pydantic
+yandex_geocoder
