+++ API/frontend/index.html
@@ -1,0 +1,24 @@
+<!DOCTYPE html>
+<html lang="en">
+
+<head>
+    <meta charset="UTF-8">
+    <meta http-equiv="X-UA-Compatible" content="IE=edge">
+    <meta name="viewport" content="width=device-width, initial-scale=1.0">
+    <title>Toilet API</title>
+</head>
+
+<body>
+    <div class="content">
+        <div>
+            <form>
+                <div>
+                    <label></label>
+                    <input placeholder="" />
+                </div>
+            </form>
+        </div>
+    </div>
+</body>
+
+</html>
