+++ MapApp/app/src/main/AndroidManifest.xml
