+++ MapApp/app/src/main/java/com/syrnnik/mapapp/MainActivity.java
