+++ MapApp/app/src/main/java/com/syrnnik/mapapp/PointsListAdapter.java
