+++ MapApp/app/src/main/java/com/syrnnik/mapapp/ServerUrls.java
