+++ MapApp/app/src/main/java/com/syrnnik/mapapp/Toilet.java
