+++ MapApp/app/src/main/java/com/syrnnik/mapapp/ToiletPoints.java
