+++ MapApp/app/src/main/java/com/syrnnik/mapapp/ToiletRouter.java
