+++ MapApp/app/src/main/res/layout/activity_add_point.xml
