+++ MapApp/app/src/main/res/layout/points_list.xml
