+++ MapApp/app/src/main/res/menu/main_menu.xml
