+++ MapApp/app/src/main/res/menu/toilet_info_menu.xml
